refactor: move rendered spec related struct into pymod
