+++ src/base/spec.rs
@@ -1,7 +1,6 @@
 use crate::prelude::*;
 
 use super::schema::{EnrichedValueType, FieldSchema};
-use pyo3::prelude::*;
 use serde::{Deserialize, Serialize};
 use std::fmt;
 use std::ops::Deref;
@@ -17,27 +16,6 @@ pub enum OutputMode {
 /// Formatting spec per output mode
 pub trait SpecFormatter {
     fn format(&self, mode: OutputMode) -> String;
-}
-
-/// A single line in the rendered spec, with optional scope and children
-#[pyclass(get_all, set_all)]
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RenderedSpecLine {
-    /// The formatted content of the line (e.g., "Import: name=documents, source=LocalFile")
-    pub content: String,
-    /// The scope name, if applicable (e.g., "documents_1" for ForEach scopes)
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub scope: Option<String>,
-    /// Child lines in the hierarchy
-    pub children: Vec<RenderedSpecLine>,
-}
-
-/// A rendered specification, grouped by sections
-#[pyclass(get_all, set_all)]
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RenderedSpec {
-    /// List of (section_name, lines) pairs
-    pub sections: Vec<(String, Vec<RenderedSpecLine>)>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
